Memory Optimization complete, Bonus?

+++ includes/fdf.h
@@ -6,7 +6,7 @@
 /*   By: m_kamal <m_kamal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 18:29:26 by mshehata          #+#    #+#             */
-/*   Updated: 2023/02/17 09:29:37 by m_kamal          ###   ########.fr       */
+/*   Updated: 2023/02/18 06:55:35 by m_kamal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,8 +35,8 @@ typedef struct s_pixel {
 typedef struct s_line {
 	t_pixel	start;
 	t_pixel	end;
-	float	dx;
-	float	dy;
+	int		dx;
+	int		dy;
 	int		decision;
 	int		color;
 }	t_line;
